refactored context
